+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 =========================
 
+* extract shared utility functions
 * remove `unpipe` package and use native `unpipe()` method
 
 2.0.1 / 2024-09-10
