+++ lib/types/json.js
@@ -13,12 +13,12 @@
  */
 
 var bytes = require('bytes')
-var contentType = require('content-type')
 var createError = require('http-errors')
 var debug = require('debug')('body-parser:json')
 var isFinished = require('on-finished').isFinished
 var read = require('../read')
 var typeis = require('type-is')
+var { getCharset, typeChecker } = require('../utils')
 
 /**
  * Module exports.
@@ -197,21 +197,6 @@ function firstchar (str) {
 }
 
 /**
- * Get the charset of a request.
- *
- * @param {object} req
- * @api private
- */
-
-function getCharset (req) {
-  try {
-    return (contentType.parse(req).parameters.charset || '').toLowerCase()
-  } catch (e) {
-    return undefined
-  }
-}
-
-/**
  * Normalize a SyntaxError for JSON.parse.
  *
  * @param {SyntaxError} error
@@ -234,17 +219,4 @@ function normalizeJsonSyntaxError (error, obj) {
   error.message = obj.message
 
   return error
-}
-
-/**
- * Get the simple type checker.
- *
- * @param {string} type
- * @return {function}
- */
-
-function typeChecker (type) {
-  return function checkType (req) {
-    return Boolean(typeis(req, type))
-  }
 }
