+++ lib/types/raw.js
@@ -15,6 +15,7 @@ var debug = require('debug')('body-parser:raw')
 var isFinished = require('on-finished').isFinished
 var read = require('../read')
 var typeis = require('type-is')
+var { typeChecker } = require('../utils')
 
 /**
  * Module exports.
@@ -87,18 +88,5 @@ function raw (options) {
       limit: limit,
       verify: verify
     })
-  }
-}
-
-/**
- * Get the simple type checker.
- *
- * @param {string} type
- * @return {function}
- */
-
-function typeChecker (type) {
-  return function checkType (req) {
-    return Boolean(typeis(req, type))
   }
 }
