+++ lib/types/text.js
@@ -11,11 +11,11 @@
  */
 
 var bytes = require('bytes')
-var contentType = require('content-type')
 var debug = require('debug')('body-parser:text')
 var isFinished = require('on-finished').isFinished
 var read = require('../read')
 var typeis = require('type-is')
+var { getCharset, typeChecker } = require('../utils')
 
 /**
  * Module exports.
@@ -92,33 +92,5 @@ function text (options) {
       limit: limit,
       verify: verify
     })
-  }
-}
-
-/**
- * Get the charset of a request.
- *
- * @param {object} req
- * @api private
- */
-
-function getCharset (req) {
-  try {
-    return (contentType.parse(req).parameters.charset || '').toLowerCase()
-  } catch (e) {
-    return undefined
-  }
-}
-
-/**
- * Get the simple type checker.
- *
- * @param {string} type
- * @return {function}
- */
-
-function typeChecker (type) {
-  return function checkType (req) {
-    return Boolean(typeis(req, type))
   }
 }
