+++ lib/types/urlencoded.js
@@ -13,13 +13,13 @@
  */
 
 var bytes = require('bytes')
-var contentType = require('content-type')
 var createError = require('http-errors')
 var debug = require('debug')('body-parser:urlencoded')
 var isFinished = require('on-finished').isFinished
 var read = require('../read')
 var typeis = require('type-is')
 var qs = require('qs')
+var { getCharset, typeChecker } = require('../utils')
 
 /**
  * Module exports.
@@ -185,21 +185,6 @@ function createQueryParser (options, extended) {
 }
 
 /**
- * Get the charset of a request.
- *
- * @param {object} req
- * @api private
- */
-
-function getCharset (req) {
-  try {
-    return (contentType.parse(req).parameters.charset || '').toLowerCase()
-  } catch (e) {
-    return undefined
-  }
-}
-
-/**
  * Count the number of parameters, stopping once limit reached
  *
  * @param {string} body
@@ -221,17 +206,4 @@ function parameterCount (body, limit) {
   }
 
   return count
-}
-
-/**
- * Get the simple type checker.
- *
- * @param {string} type
- * @return {function}
- */
-
-function typeChecker (type) {
-  return function checkType (req) {
-    return Boolean(typeis(req, type))
-  }
 }
